pyservice/requirements.txt
fastapi
black
uvicorn
pydantic
jinja2
pdfkit
playwright
playwright-stealth
google-adk
litellm